% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/back.R
\name{back}
\alias{back}
\title{Backup variable during development}
\usage{
back(x)
}
\arguments{
\item{x}{\itemize{
\item any variable supported by the qs package
}}
}
\value{
The variable itself is stored in the directory "back". If there is none, it creates it in the script folder.
}
\description{
Backup variable during development
}
\examples{
\dontrun{
x <- 1:100
back(x)
}
}
