% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChecDateNull.R
\name{ChecDateNull}
\alias{ChecDateNull}
\title{Empty table}
\usage{
ChecDateNull(x)
}
\arguments{
\item{x}{\itemize{
\item table
}}
}
\value{
\itemize{
\item stops the script if the table is empty
}
}
\description{
Empty table
}
\examples{
\dontrun{
df <- iris
ChecDateNull(df)
df <- df[df$Species=="any"]
ChecDateNull(df)
}

}
