% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChecFile.R
\name{ChecFile}
\alias{ChecFile}
\title{File Existence}
\usage{
ChecFile(file_path)
}
\arguments{
\item{file_path}{\itemize{
\item file name
}}
}
\value{
Stops the script if there is no file
}
\description{
File Existence
}
