% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChecFile_sql.R
\name{ChecFile_sql}
\alias{ChecFile_sql}
\title{File request Existence}
\usage{
ChecFile_sql(file.sql, folder.sql = "sql")
}
\arguments{
\item{file.sql}{\itemize{
\item Request file
}}

\item{folder.sql}{\itemize{
\item Request folder. Default "sql"
}}
}
\value{
Stops the script if there is no file
}
\description{
File request Existence
}
\examples{
\dontrun{
file <- tempfile()
ChecFile(file)
file <- "any.rds"
ChecFile(file)
}

}
