% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChecFileDateDay.R
\name{ChecFileDateDay}
\alias{ChecFileDateDay}
\title{Checks the creation date of a file}
\usage{
ChecFileDateDay(file_path, day.ch = Sys.Date())
}
\arguments{
\item{file_path}{-file name}

\item{day.ch}{\itemize{
\item The date the file was created. The default is today.
}}
}
\value{
Stops the script on a date mismatch
}
\description{
Checks the creation date of a file
}
