% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChecFileDateWeek.R
\name{ChecFileDateWeek}
\alias{ChecFileDateWeek}
\title{Check if the file creation date matches the current week}
\usage{
ChecFileDateWeek(file_path)
}
\arguments{
\item{file_path}{-file name}
}
\value{
Stops the script on a date mismatch
}
\description{
Check if the file creation date matches the current week
}
