% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChecNotNull.R
\name{ChecNotNull}
\alias{ChecNotNull}
\title{Checks if the value of a variable is not equal to 0}
\usage{
ChecNotNull(x)
}
\arguments{
\item{x}{\itemize{
\item any variable
}}
}
\value{
stops the script with an error if the variable is not equal to 0
}
\description{
Checks if the value of a variable is not equal to 0
}
\examples{

x <- 0
try(ChecNotNull(x))


x <- 2
try(ChecNotNull(x))

}
