% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChecNull.R
\name{ChecNull}
\alias{ChecNull}
\title{Checks if the value of a variable is equal to 0}
\usage{
ChecNull(x)
}
\arguments{
\item{x}{\itemize{
\item any variable
}}
}
\value{
stops the script with an error if the variable is equal to 0
}
\description{
Checks if the value of a variable is equal to 0
}
\examples{
x <- 0
try(ChecNotNull(x))

x <- 2
try(ChecNotNull(x))

}
