% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChecVar.R
\name{ChecVar}
\alias{ChecVar}
\title{Missing variable}
\usage{
ChecVar(x)
}
\arguments{
\item{x}{\itemize{
\item any variable
}}
}
\value{
\itemize{
\item stops the script if no variable is found
}
}
\description{
Missing variable
}
\examples{
x <- 1:100
try(ChecVar(x))

rm(x)
try(ChecVar(x))

}
