% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjDT.R
\name{cjDT}
\alias{cjDT}
\title{Cartesian join for two data.table}
\usage{
cjDT(x, y)
}
\arguments{
\item{x}{\itemize{
\item data.table
}}

\item{y}{\itemize{
\item data.table
}}
}
\value{
Cartesian join for two data.table
}
\description{
Cartesian join for two data.table
}
