% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecuteDT.R
\name{ExecuteDT}
\alias{ExecuteDT}
\title{Execute an update statement, query number of rows affected, and then close result set}
\usage{
ExecuteDT(con, file.sql, folder.sql = "sql")
}
\arguments{
\item{con}{\itemize{
\item Connect
}}

\item{file.sql}{\itemize{
\item Request file
}}

\item{folder.sql}{\itemize{
\item Request folder. Default "sql"
}}
}
\value{
Same as DBI::dbExecute but with logging and request in another file
}
\description{
Execute an update statement, query number of rows affected, and then close result set
}
