% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_run.R
\name{log_run}
\alias{log_run}
\title{Creating and running a log file}
\usage{
log_run(log_name)
}
\arguments{
\item{log_name}{The name of the log file. Better if it matches the name of the script}
}
\value{
If not, it creates a log folder, after which it organizes the file
}
\description{
Creating and running a log file
}
