% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_proj_fold.R
\name{new_proj_fold}
\alias{new_proj_fold}
\title{Create project folders}
\usage{
new_proj_fold(
  fold = c("archive", "back", "date", "log", "materials", "py", "result", "scripts",
    "sql")
)
}
\arguments{
\item{fold}{\itemize{
\item A vector with the names of folders that should be in the project.
Default:
\item archive - Folder for saving old versions
\item back - Backup folder
\item date - Inbox folder
\item log - Log folder
\item materials - Supporting data, descriptions, notes
\item py - Python code
\item result -
\item scripts - Supporting scripts, functions, settings
\item sql - SQL code.
}}
}
\value{
Creates folders
}
\description{
Create project folders
}
