% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryDT.R
\name{QueryDT}
\alias{QueryDT}
\title{Send query, retrieve results and then clear result set}
\usage{
QueryDT(con, file.sql, folder.sql = "sql")
}
\arguments{
\item{con}{\itemize{
\item Connect
}}

\item{file.sql}{\itemize{
\item Request file
}}

\item{folder.sql}{\itemize{
\item Request folder. Default "sql"
}}
}
\value{
Same as DBI::dbGetQuery but with logging and returns data.table
}
\description{
Send query, retrieve results and then clear result set
}
