% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryDT_S.R
\name{QueryDT_S}
\alias{QueryDT_S}
\title{Send query, retrieve results and then clear result set}
\usage{
QueryDT_S(con, select, conv = FALSE, enc.conv = "CP1251")
}
\arguments{
\item{con}{\itemize{
\item Connect
}}

\item{select}{\itemize{
\item Request select
}}

\item{conv}{\itemize{
\item Convert result
}}

\item{enc.conv}{\itemize{
\item Encoding from which
}}
}
\value{
Same as DBI::dbGetQuery but with logging and returns data.table
}
\description{
Send query, retrieve results and then clear result set
}
