% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dtv.R
\name{read_dtv}
\alias{read_dtv}
\title{Read a UTF-8 encoded text file downloaded from. Package wrapper vroom.}
\usage{
read_dtv(
  file_path,
  skp = 0,
  gm = 1000,
  head_nm_fl = NA,
  nm_old = "old",
  nm_new = "new",
  nm_type_var = "type_var",
  type_var = FALSE
)
}
\arguments{
\item{file_path}{-Path to file}

\item{skp}{\itemize{
\item How many lines to skip. Default 0.
}}

\item{gm}{\itemize{
\item How many rows to take if there is no file with column name encoding, or the variable types in this file are not specified. The default is 1000.
}}

\item{head_nm_fl}{\itemize{
\item Path to a file with input and output variable names and variable types (optional). Uses package openxlsx.
}}

\item{nm_old}{\itemize{
\item Field with the names of the input columns. In the file head_nm_fl.
}}

\item{nm_new}{-Field with the names of the output columns. In the file head_nm_fl.}

\item{nm_type_var}{\itemize{
\item Field with variable type designations. In the file file head_nm_fl.s
}}

\item{type_var}{\itemize{
\item Specifies whether there is a field with variable types. Defaults to FALSE.
}}
}
\value{
returns data.table
}
\description{
Read a UTF-8 encoded text file downloaded from. Package wrapper vroom.
}
