% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{read_sql}
\alias{read_sql}
\title{Query text with the substituted parameters}
\usage{
read_sql(file.sql, folder.sql = "sql")
}
\arguments{
\item{file.sql}{\itemize{
\item Request file
}}

\item{folder.sql}{\itemize{
\item Request folder. Default "sql"
}}
}
\value{
\itemize{
\item Returns the query text with the substituted parameters
}
}
\description{
Query text with the substituted parameters
}
