% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_sleep.R
\name{sys_sleep}
\alias{sys_sleep}
\title{Sys.sleep alter}
\usage{
sys_sleep(val, unit = c("s", "ms", "us", "ns"))
}
\arguments{
\item{val}{\itemize{
\item time delay
}}

\item{unit}{\itemize{
\item units of time
}}
}
\value{
\itemize{
\item absent
}
}
\description{
Sys.sleep alter
}
